requirements.txt
# qalembic==1.14.0
bleach==6.2.0
blinker==1.9.0
Bootstrap-Flask==2.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
dominate==2.9.1
email_validator==2.2.0
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==1.0.0
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-SocketIO==5.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
SQLAlchemy==2.0.36
typing_extensions==4.12.2
visitor==0.1.3
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
gunicorn==21.2.0
# psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests>=2.28.1
Flask-Dance==5.0.0